requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.1
SQLAlchemy==2.0.25
python-dotenv==1.0.1
bcrypt==4.1.2
supabase==2.3.4
python-jose[cryptography]==3.3.0
httpx==0.26.0
postgrest==0.13.1
